pyproject.toml
[project]
name = "the-flask-mega-tutorial"
version = "0.1.0"
description = ""
authors = [
    {name = "Timur Berezin",email = "corporate.timur01@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "flask (>=3.1.0,<4.0.0)",
    "pytest (>=8.3.4,<9.0.0)",
    "flask-wtf (>=1.2.2,<2.0.0)",
    "flask-migrate (>=4.1.0,<5.0.0)",
    "flask-login (>=0.6.3,<0.7.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "aiosmtpd (>=1.4.6,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "flask-mail (>=0.10.0,<0.11.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "flask-moment (>=1.0.6,<2.0.0)",
    "flask-babel (>=4.0.0,<5.0.0)",
    "langdetect (>=1.0.9,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "whoosh (>=2.7.4,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
